+++ src/cortex.js
@@ -28,9 +28,24 @@ Cortex = (function(_super, _cortexPubSub) {
 
   __extends(Cortex, _super);
 
-  Cortex.prototype.on = function(event, callback) {
-    if(event === "update") {
+  Cortex.prototype.on = function(eventName, callback) {
+    if(eventName === "update") {
       this.__callbacks.push(callback);
+    }
+  };
+
+  Cortex.prototype.off = function(eventName, callback) {
+    if(eventName === "update") {
+      if(callback) {
+        for(var i=0, ii=this.__callbacks.length;i < ii;i++) {
+          if(callback === this.__callbacks[i]) {
+            this.__callbacks.splice(i, 1);
+            break;
+          }
+        }
+      } else {
+        this.__callbacks = [];
+      }
     }
   };
 
@@ -42,13 +57,13 @@ Cortex = (function(_super, _cortexPubSub) {
     this.__setValue(newValue, path);
     this.__rewrap(path);
 
-    this.__callbacks.forEach(function (callback) {
-        if(callback) {
-            callback(this);
-        }
-    });
-    return true;
+    for(var i=0, ii=this.__callbacks.length;i < ii;i++) {
+      if(this.__callbacks[i]) {
+        this.__callbacks[i](this);
+      }
+    }
 
+    return true;
   };
 
   Cortex.prototype.__subscribe = function() {