+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.5.3 (May 6, 2014)
+Features:
+- Support multiple callbacks
+- Add ``.off("update", callback)`` method
+
 # 0.5.2 (May 2, 2014)
 
 Features: